+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 - Added grid state management functions `getGridState()` and `setGridState()` to the `Table.View` class to allow extensions to save and restore the complete state of the table view, including column positions, sorting, filtering, and other grid configurations. [#3751](https://github.com/zowe/zowe-explorer-vscode/pull/3751)
   New `ProfilesCache.isCertFileValid()` method to check validity of profile's certificate file if certificate auth used, returning a boolean value. [#3699](https://github.com/zowe/zowe-explorer-vscode/issues/3699)
 - Changed `request` property in the `PollRequest` type to now support both asynchronous and synchronous functions. [#3764](https://github.com/zowe/zowe-explorer-vscode/pull/3764/)
+- Added new getCount API to provide ability to return total number of datasets for a given filter.[#3844](https://github.com/zowe/zowe-explorer-vscode/pull/3844)
 
 ### Bug fixes